modularized reducer files for easier reading
